upbit_bot/requirements.txt
requests
PyJWT
numpy
pyyaml
pandas
python-dateutil
pyupbit
pandas_ta
